API

MakieThemes.BBC.theme_bbc

MakieThemes.demofigure

MakieThemes.demoscatter

MakieThemes.demofigure — Function

demofigure([theme::Attributes])

Shows off a demonstration figure with the given theme (defaults to the current theme). The figure contains scatters, surfaces, bands, and barplots, as well as legends, titles, etc. It is able to show off most aspects of the theme!

MakieThemes.demoscatter — Function

demoscatter([theme::Attributes]; legend_vertical = true, legend_outside_axis = true)

Shows off a demonstration scatter plot with the given theme (defaults to the current theme). A legend is also plotted.